Separate ParseArgumentVisitor
